feat: Update pygls to v0.13.0
For some reason the new version of YLS did not work on vscode v1.73.1 .
It was weird because we send a message a but did not get anything back.
I don't think it was caused by my current setup. So I tried to update to
a newer version of pygls which seem to help.